GPU ablation sweep: wandb, mmap data pipeline, bf16+flash, SIGReg fixes
Major changes:
- wandb integration with deferred init and aggregated metric logging
- Rewrote data pipeline: one-time tokenization to flat .npy files with
  EOT/BOS document framing and pre-shuffled windows (nanochat pattern)
- bf16 mixed precision via torch.autocast, enabling FlashAttention2 kernel
  through SDPA is_causal=True (bypasses 4D mask when no padding)
- Fixed SIGReg to regularize token embeddings, not prediction vectors
- Added pred_loss_scale to balance L2 prediction vs SIGReg contribution
- Added use_prediction_head toggle — backbone output directly as prediction
- Best recipe (gpu_l2_5k_nohead): top1=24%, median_rank=191 at 5k steps,
  both losses alive and converging. emb_init=1.0, λ=0.05, pred_scale=10.

+++ src/eval.py
@@ -39,8 +39,8 @@ def compute_retrieval_metrics(
     valid_targets = target_ids[target_mask]
     if valid_targets.numel() == 0:
         raise RuntimeError("No valid targets available for retrieval metrics.")
-    valid_prediction = prediction[target_mask]
-    embedding_table = model.token_embeddings.weight
+    valid_prediction = prediction[target_mask].float()
+    embedding_table = model.token_embeddings.weight  # already fp32
 
     ranks: list[torch.Tensor] = []
     top_hits = {1: 0, 5: 0, 10: 0}
@@ -89,7 +89,7 @@ def compute_ce_nll(
     target_ids: torch.Tensor,
     target_mask: torch.Tensor,
 ) -> tuple[float, int]:
-    logits = prediction[target_mask]
+    logits = prediction[target_mask].float()
     targets = target_ids[target_mask]
     nll = torch.nn.functional.cross_entropy(logits, targets, reduction="sum")
     return float(nll.item()), int(targets.numel())
@@ -105,8 +105,8 @@ def compute_harmax_nll(
     exponent: float | None = None,
 ) -> tuple[float, int]:
     valid_targets = target_ids[target_mask]
-    valid_prediction = prediction[target_mask]
-    embedding_table = model.token_embeddings.weight
+    valid_prediction = prediction[target_mask].float()
+    embedding_table = model.token_embeddings.weight  # already fp32
     pow_n = float(exponent if exponent is not None else model.config.model.d_model)
     total_nll = 0.0
     total_tokens = 0
